+++ src/Controller/ParticipantController.php
@@ -7,6 +7,7 @@ use App\Entity\Search;
 use App\Form\ParticipantFormType;
 use App\Form\SearchFormType;
 use App\Repository\ParticipantRepository;
+use App\Repository\UserEntityRepository;
 use App\Services\SimpleSearchService;
 use App\Services\PaginatorService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -18,19 +19,20 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/participant', name: 'participant_')]
 class ParticipantController extends AbstractController
 {
+    
     #[Route('s/{pagina}', defaults: ['pagina' => 1], name: 'list')]
     public function list(
                 int $pagina,
                 Request $request,
                 PaginatorService $paginator,
-                SimpleSearchService $searchService
+                SimpleSearchService $searchService,
+                ParticipantRepository $participantRepository
                 ):Response 
     {
+        $this->denyAccessUnlessGranted('see', (new Participant));
         $busqueda = new Search();
         $busqueda->setEntity(Participant::class);
-
         $busqueda = $searchService->getSearchFromSession(Participant::class) ?? $busqueda;
-
         $searchForm = $this->createForm(SearchFormType::class, $busqueda, [
             'field_choices' => [
                 'Nom' => 'name',
@@ -41,21 +43,15 @@ class ParticipantController extends AbstractController
                 'Nom' => 'name',
             ]
         ]);
-
         $searchForm->handleRequest($request);
         $searchService->setSearch($busqueda);
-
-    
         $participants = $paginator->paginate(
-            $searchService->prepareQuery(),
+            $participantRepository->participantsPerEntity($busqueda),
             $pagina
         );
-
         $searchService->storeSearchInSession($busqueda);
-
         if($searchForm->isSubmitted() && $searchForm->isValid())
             return $this->redirectToRoute('participant_list');
-    
         return $this->renderForm('participant/list.html.twig', [
             'search' => $searchForm,
             'paginator' => $paginator,
@@ -63,27 +59,26 @@ class ParticipantController extends AbstractController
         ]);
     }
 
-
     
     #[Route('/crear', name: 'create')]
     public function create(
                 Request $request,
-                ParticipantRepository $participantRepository
+                ParticipantRepository $participantRepository,
+                UserEntityRepository $userEntityRepository,
             ): Response
     {
         $participant = new Participant();
-
+        $this->denyAccessUnlessGranted('create', $participant);
+        $user_entity_id = $request->getSession()->get('user_entity_id');
+        $entity = $userEntityRepository->find($user_entity_id)->getEntity();
         $form = $this->createForm(ParticipantFormType::class, $participant);
         $form->handleRequest($request);
-        
         if($form->isSubmitted() && $form->isValid()) {
-
             $this->addFlash('success', "Participant ". $participant->getName() ." amb éxit.");
+            $participant->setEntity($entity);
             $participantRepository->add($participant, true);
-
             return $this->redirectToRoute('participant_list');
         }
-
         return $this->renderForm('participant/create.html.twig', [
             'formulario' => $form
         ]);
@@ -97,47 +92,41 @@ class ParticipantController extends AbstractController
                 ParticipantRepository $participantRepository
             ): Response
     {
-
+        $this->denyAccessUnlessGranted('edit', $participant);
         $form = $this->createForm(ParticipantFormType::class, $participant);
         $form->handleRequest($request);
-        
         if($form->isSubmitted() && $form->isValid()) {
-
             $this->addFlash('success', "Participant ". $participant->getName() ." editat amb éxit.");
             $participantRepository->add($participant, true);
-
             return $this->redirectToRoute('participant_list');
         }
-
         return $this->renderForm('participant/edit.html.twig', [
             'formulario' => $form,
             'participant' => $participant
         ]);
     }
 
-
     
     #[Route('/create/tiquet', name: 'create_with_ticket')]
     public function createWithTiquet(
                 Request $request,
-                ParticipantRepository $participantRepository
+                ParticipantRepository $participantRepository,
+                UserEntityRepository $userEntityRepository,
             ): Response
     {
         $participant = new Participant();
-        
+        $this->denyAccessUnlessGranted('create', $participant);
         $activity_id = $request->request->get('activity_id');
-        
+        $user_entity_id = $request->getSession()->get('user_entity_id');
+        $entity = $userEntityRepository->find($user_entity_id)->getEntity();
         $form = $this->createForm(ParticipantFormType::class, $participant);
         $form->handleRequest($request);
-        
         if($form->isSubmitted() && $form->isValid()) {
-            
+            $participant->setEntity($entity);
             $participantRepository->add($participant, true);
-          
             return $this->redirectToRoute('ticket_create_in_activity_after_participant', ['participant' => $participant->getId(), 'activity' => (int) $activity_id]);
         }
     }
-
 
 
     #[Route('/eliminar/{id}', name: 'delete')]
@@ -146,16 +135,14 @@ class ParticipantController extends AbstractController
                 ParticipantRepository $participantRepository
             ): Response
     {
-
+        $this->denyAccessUnlessGranted('delete', $participant);
         $this->addFlash('success', "El participant " . $participant->getName() . " s'ha eliminat correctament.");
         $participantRepository->remove($participant, true);
-        
         return $this->redirectToRoute('participant_list');
     }
 
-      /**
-     * @Route("/search/forget", name="forget_search")
-     */     
+
+    #[Route('/search/forget', name: 'forget_search')]    
     public function forgetSearch(
         SimpleSearchService $searchService
         ): Response
